+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "python.pythonPath": "V:\\anaconda3\\python.exe"
+}
