+++ processor/Dockerfile
@@ -1,0 +1,10 @@
+FROM python:3
+
+# [Optional] If your pip requirements rarely change, uncomment this section to add them to the image.
+COPY requirements.txt /tmp/pip-tmp/
+RUN pip3 --disable-pip-version-check --no-cache-dir install -r /tmp/pip-tmp/requirements.txt \
+    && rm -rf /tmp/pip-tmp
+
+COPY . /processor_app
+WORKDIR /processor_app
+CMD python can_converter.py
