--- requirements.txt
@@ -1,4 +1,0 @@
-pandas
-numpy
-asammdf
-flask
